+++ releasenotes/notes/gevent-warning-f91ed0e6cac71a21.yaml
@@ -1,0 +1,5 @@
+---
+features:
+  - |
+    A warning has been added to alert when gevent monkey patching is done after
+    ddtrace has been imported.
