+++ tests/contrib/gevent/wrong_program_gevent.py
@@ -1,6 +1,6 @@
 from gevent import monkey
 
-import ddtrace.profiling.auto  # noqa
+import ddtrace  # noqa
 
 
 monkey.patch_all()
